Implement OptionMapper insert, update and build
See TODO in OptionMapper.buildObjectFromResultSet:
Should we call PreguntaMapper.findById to set the referenced Question,
or should we return an incomplete Opcion object?

Note: the schema has changed — as we must be able to change the question
order, we can't have that field as part of a primary key. We also can't
make `questionId` the only key, as it is our foreign key.

Therefore, we have to create a surrogate auto increment key on answer,
remove questionOrder from the key, and continue using questionId as our
foreign key. The `user_questions` had to change to reflect that.