Add unique constraint on answer schema
The pair (questionId, questionOrder) in now unique. Any attempt to
insert duplicate questions would result in an exception.

However, we now must ensure to balance this limitation correctly when
swapping questions. Since we can't have duplicates, we can't just swap
the order of the 1st and 2nd question. We would have to do 1->0, 2->1,
0->2 for that.