Remove mention of `answerId` from OptionMapper
Even though `answerId` is the real id of the `answers` table, using it
as the id inside the application is cumbersome, as the Option object has
no notion of an identifier. Therefore, we use {questionId,
questionOrder} as the id so we can update / delete without knowing the
real numerical value of the answer.